feat : PR 반영 (User-tagCode 도 Long 으로 )

+++ src/main/java/com/archiservice/user/domain/User.java
@@ -27,5 +27,5 @@ public class User {
     private LocalDateTime updatedAt;
 
     private String ageCode;
-    private BigInteger tagCode;
+    private Long tagCode;
 }
